build: remove the need to use environment variable when building the player, demo or tests
