+++ package.json
@@ -25,15 +25,15 @@
   "license": "Apache-2.0",
   "sideEffects": false,
   "scripts": {
-    "build": "webpack --progress --config webpack.config.js",
+    "build": "webpack --progress --config webpack.config.js --env production",
     "build:all": "npm run build && npm run build:min && npm run build:wasm:release && npm run build:modular",
-    "build:min": "cross-env RXP_MINIFY=true webpack --progress --config webpack.config.js",
-    "build:min:report": "cross-env RXP_SIZE_REPORT=true cross-env RXP_MINIFY=true webpack --progress --config webpack.config.js",
+    "build:min": "webpack --progress --config webpack.config.js --env minify --env production",
+    "build:min:report": "webpack --progress --config webpack.config.js --env minify --env production --env reportSize",
     "build:modular": "./scripts/build/generate_build.js",
-    "build:report": "cross-env RXP_SIZE_REPORT=true webpack --progress --config webpack.config.js",
+    "build:report": "webpack --progress --config webpack.config.js --env production --env reportSize",
     "build:rxp:all": "npm run build && npm run build:min && npm run build:modular",
-    "build:watch": "webpack --progress --config webpack.config.js -w",
-    "build:min:watch": "cross-env RXP_MINIFY=true webpack --progress --config webpack.config.js -w",
+    "build:watch": "webpack --progress --config webpack.config.js -w -env production",
+    "build:min:watch": "webpack --progress --config webpack.config.js -w --env production --env minify",
     "build:wasm:debug": "cd ./src/parsers/manifest/dash/wasm-parser && cargo build --target wasm32-unknown-unknown && cp target/wasm32-unknown-unknown/debug/mpd_node_parser.wasm ../../../../../dist/mpd-parser.wasm",
     "build:wasm:release": "cd ./src/parsers/manifest/dash/wasm-parser && cargo build --target wasm32-unknown-unknown --release && wasm-opt -O3 -o ../../../../../dist/mpd-parser.wasm target/wasm32-unknown-unknown/release/mpd_node_parser.wasm && cd ../../../../../ && npm run wasm-strip",
     "certificate": "./scripts/generate_certificate",
@@ -42,18 +42,18 @@
     "check:appveyor": "npm run check:types && npm run lint && npm run lint:demo && npm run lint:tests && npm run test:appveyor",
     "check:travis": "npm run check:types && npm run lint && npm run lint:demo && npm run lint:tests && npm run test:travis",
     "check:types": "tsc --noEmit --project .",
-    "demo": "cross-env RXP_METAPLAYLIST=true node ./scripts/generate_full_demo.js",
-    "demo:min": "cross-env RXP_METAPLAYLIST=true RXP_MINIFY=true node ./scripts/generate_full_demo.js",
-    "demo:watch": "cross-env RXP_METAPLAYLIST=true node ./scripts/generate_full_demo.js --watch",
+    "check:types:watch": "tsc --noEmit --watch --project .",
+    "demo": "node ./scripts/generate_full_demo.js --production-mode",
+    "demo:min": "node ./scripts/generate_full_demo.js --production-mode --minify",
+    "demo:watch": "node ./scripts/generate_full_demo.js --watch --production-mode",
     "doc": "rm -rf doc/generated; node ./scripts/doc-generator/index.js doc/ doc/generated",
     "lint": "eslint src -c .eslintrc.js",
     "lint:demo": "eslint demo/full/scripts",
     "lint:tests": "eslint tests/**/*.js",
     "list": "node scripts/list-npm-scripts.js",
     "prepublishOnly": "npm run build:modular",
-    "standalone": "cross-env RXP_METAPLAYLIST=true RXP_ENV=development node ./scripts/run_standalone_demo.js",
-    "start": "cross-env RXP_METAPLAYLIST=true RXP_ENV=development node ./scripts/run_full_demo.js",
-    "start:min": "cross-env RXP_METAPLAYLIST=true RXP_MINIFY=true node ./scripts/run_full_demo.js",
+    "standalone": "node ./scripts/run_standalone_demo.js",
+    "start": "node ./scripts/start_demo_web_server.js",
     "wasm-strip": "node scripts/wasm-strip.js dist/mpd-parser.wasm",
     "test:appveyor": "npm run test:unit && npm run test:memory",
     "test:integration": "node tests/integration/run.js --bchromehl --bfirefoxhl",
@@ -99,7 +99,6 @@
     "chai": "4.3.4",
     "cheerio": "1.0.0-rc.10",
     "core-js": "3.17.1",
-    "cross-env": "7.0.3",
     "eslint": "7.32.0",
     "eslint-plugin-import": "2.24.2",
     "eslint-plugin-jsdoc": "36.0.8",
@@ -139,8 +138,7 @@
   },
   "scripts-list": {
     "Build a demo page (e.g. to test a code change)": {
-      "start": "Build and launch the \"full\" demo (with a UI) with the non-minified RxPlayer on a local server. Re-build on file updates.",
-      "start:min": "Build and launch the \"full\" demo (with a UI) with the minified RxPlayer on a local server. Re-build on file updates.",
+      "start": "Build the \"full\" demo (with a UI) with the non-minified RxPlayer and serve it on a local server. Re-build on file updates.",
       "demo": "Build the demo in demo/bundle.js",
       "demo:min": "Build the demo and minify it in demo/bundle.js",
       "demo:watch": "Build the demo in demo/bundle.js each times the files update.",
