+++ tests/memory/run.js
@@ -5,7 +5,9 @@ const karma = require("karma");
 const parseConfig = karma.config.parseConfig;
 const Server = karma.Server;
 const TestContentServer = require("../contents/server");
-const webpackConfig = require("../../webpack-tests.config.js");
+const generateTestWebpackConfig = require("../generate_test_webpack_config");
+
+const CONTENT_SERVER_PORT = 3000;
 
 const argv = process.argv;
 if (argv.includes("-h") || argv.includes("--help")) {
@@ -13,9 +15,7 @@ if (argv.includes("-h") || argv.includes("--help")) {
   process.exit(0);
 }
 
-const singleRun = argv.includes("--watch") ?
-  false :
-  !process.env.RXP_TESTS_WATCH;
+const singleRun = !argv.includes("--watch");
 
 const browsers = [];
 if (argv.includes("--bchrome")) {
@@ -29,6 +29,10 @@ if (browsers.length === 0) {
   displayHelp();
   process.exit(0);
 }
+
+const webpackConfig = generateTestWebpackConfig({
+  contentServerInfo: { url: "127.0.0.1", port: CONTENT_SERVER_PORT },
+});
 
 const karmaConf = {
   basePath: "",
@@ -73,7 +77,7 @@ const karmaConf = {
   },
 };
 
-const testContentServer = TestContentServer(3000);
+const testContentServer = TestContentServer(CONTENT_SERVER_PORT);
 parseConfig(
   null,
   karmaConf,
